requirements.txt
# Options Scanner Dependencies
# Install: pip install -r requirements.txt

# Market Data
yfinance>=0.2.33
pandas>=2.0.0
numpy>=1.24.0

# News Scraping
beautifulsoup4>=4.12.0
feedparser>=6.0.0
requests>=2.31.0
lxml>=5.0.0

# Terminal UI
rich>=13.0.0

# Web App
flask>=3.0.0
flask-cors>=4.0.0

# Config
python-dotenv>=1.0.0
